Interpolation of CLAMS age to profile eqlat/theta now better behaved
I replaced the interp2d interpolator in tccon_priors.get_mean_age with
a LinearNDInterpolator. This avoids the extremely noisy age that varies
by many years from level to level and also correctly does not try to
fill values outside the range of eqlat/theta covered by the CLAMS
model. We now fill the CO2 above the top where CLAMS can cover with
the top non-NaN value.

Also the age lookup - had a simple sign error before causing it to look
to the future for stratospheric CO2. This necessitated fixing the MLO/SMO
record to fill in missing months. All the lookup functions assume that
there is data for each month, this is now true. The CO2 records now
include an "interp_flag" column that indicates if that month had to be
filled by interpolation.