HF LUT updated, HF calculation revised
Key change is that the HF derived from the slopes now uses point
by point specific strat. bdy. condition rather than an annual
average. This eliminates an issue where, if you look at a time
series of HF for a given age and theta in the HFTropicsRecord.conc_strat,
it as a very sawtooth-like pattern as the CH4 concentration drops below
the annual average.

Also changed the HF calculation (in _calc_age_spec_gas) to avoid
calculating the same dates multiple times - the loop is now over
years instead of dates.