F(gas) now applied separately from the convolution
Until now, we've included chemical loss in the stratosphere by
modifying the age spectra with the fraction of gas remaining,
F(gas), so that the contribution of the strat. bdy. condition
from a given age is reduced by the fraction lost for that age.

However, the F(gas) values were calculated from ACE profiles,
not for a "pure" age, but for air parcels that are the result
of convolving the age spectra with the concentrations. Therefore,
the F(gas) values should *not* be multiplied by the age spectra
because effectively they already are.