Refactored ACE val and the aircraft/aircore val
ACE val - pulled out some of the matching logic in hopes of reusing
it with aircraft/aircore. Didn't happen that way, but possibly useful
in the future.

Aircraft/aircore val - modified a lot of the methods that rely on
iteration to allow me to pass the map and atm directories directly
or by type. Hopefully this will be a little cleaner in the long
run. Also added method to return a profile-by-profile binned array.

Preprocessing - added function to write out all the date/lat/lons
that match the .atm files.