Address some issues with extending the obs. profile
Files modified: ginput/priors/backend_analysis/retrieval_obs_val.py

In case the obs. profile in an .atm file only goes up to the ceiling
(hasn't been extended with an old prior) we need to append new levels
so that the alt-weighted bins can smoothly blend from the top of the
obs. profile to the prior. There were some issues with the initial
implementation that have been corrected.