Fixed issues with matching .atm and .map files in aircraft_aircore_val
Two fixes:

1. Write lat/lon with much greater precision to the list of profiles files.
This helps mollify the underlying issue, which was that when mod_maker calculates
the locations dictionary, it adds and subtracts 360 degrees several times
in order to make sure it's using the right convention for longitude. With
only 3 decimals of precision, this can lead to floating point errors that
change how it rounds to 2 decimal places. Adding more precision to the list
buries the floating point error.

2. Then, in the data iterators, I could stop rounding to 3 decimal places
before figuring out the .map subdirectory but have to round to 2 before
formatting the .map file name.