Bugfix in aircraft-prior stitching
When adjusting the profile to extend the top observations to the
tropopause then interpolate to the overworld in theta-space, the
case where the observations actually ended in the middle world was
not handled correctly. It would overwrite the observations in the
middleworld with an interpolation between the top obs. value and
the overworld prior. That's clearly not ideal, so it now only
interpolates altitudes above the top of the observations.