Added proper .map file writer
The new .map file writer takes a .vmr and .mod file as input and generates
a .map file with the primary gases on the same vertical grid as the .vmr
file. It can write either the old-style text files or new-style netCDF4
files. The latter have been verified to be CF-1.7 compliant with
cfchecks (https://github.com/cedadev/cf-checker).