writers.py now has option to skip CF unit compliance if import failed
I've had trouble with the CFUnits package failing to import with an
assertion error, which looks like it links to the wrong C library
or something. This problem happens most frequently when importing
in Jupyter Notebooks, even after a clean install of the environment.

To get around this, if CFUnits fails to import, the writers that rely
on it can be given `no_cfunits=True` to skip enforcing CF-compliance.