Added ability to generate .mod files from a runlog
Still testing but the basic behavior is there. run_ginput.py
has a new pair of subcommands (`rlmod` and `tccon-rlmod`) that
read a given runlog, figure out unique date ranges, lats, lons,
and sites in the runlog, and call the mod_maker driver function
for each of those.

The main limitation at the moment is that each unique site/lat/lon
combination *must* have a single obs. alt, simply because there is
no mechanism to write separate files when the site/lat/lon are all
the same but the alt is different (the files will have the same names.)
There is a check, and it will print a warning, but it is pretty simple-
minded and only checks that the lat/lon are numerically identical,
it does *not* check if, after rounding, they will be the same in the
output file name.