Modified `MloSmoTropicsRecord` to allow truncating the MLO/SMO data at a certain date
This allowed modification of the `acos_interface` to by default truncate
the record to the month that the granule is in. It was also modified to
write the MLO/SMO seasonal record to the output .h5 file.

`MloSmoTropicsRecord` also now forces stratosphere LUT recalculation if
custom start, end, or truncation dates are specified, but this can be
overridded. That enables the HF record to instantiate a CH4 record with
specified start and end dates without recalculating the LUT unless necessary.

The `mlo_smo_prep` module had some unused code removed and docstrings added.