Priors code now allows NaNs to fullfill the truncate date check
Previously if the input MLO/SMO files had NaNs at then end of their
record because of e.g. instrument failure then those months would
not count when the MloSmoTraceGasRecord instance checked if the MLO/SMO
data reached a late enough month to satisfy the requirement that it
reach the truncation date. While this makes some sense, it meant
that data that would always be missing would prevent the priors
code from running until a later month was filled in, which introduces
an unacceptable potential delay.

The acos_interface was also updated slightly to check that the
paths to MLO and SMO files provided actually exist before starting
any computationally intensive calculations.