Fixed conda environment cfunits constraint, updated versions in test .vmr files
Updating the cfunits version constraint in the setup.py file isn't sufficient,
I also need to update it in the environment.yml file. The new .yml file works
for a fresh install; however there is a deprecation warning that seems to
arise from a version disconnect between netCDF4 and something else (maybe numpy).

I did try doing a completely fresh install (just specifying which packages and
not which versions) but this didn't work - pydap installed with Python 3.10 tried
to use `collections.Mapping` which doesn't exist in Python 3.10.