Handle case in `mlo_smo_prep` when all data is flagged
If all the hourly NOAA data being used by `mlo_smo_prep` to update the monthly
file is flagged, fill values, or otherwise removed before the call to
`_noaa_prelim_flagging`, then the hour-to-hour difference filter can fail when
it tries to copy the flag values for the first and last data points. This commit
introduces a check that skips that step if there is no data to filter.