Updated priors code to scale GEOS IT CO differently
This does change the `.mod` and `.vmr` file headers slightly, both
now include an extra line to mark the source of CO. For the `.mod`
file, this is essential so that the VMR generation knows whether or
not to scale the tropospheric CO. For the `.vmr` files, this is
strictly to improve tracability.

I've also gone ahead and bumped the `ginput` version to 1.2.0 but not
tagged this commit yet. My reasoning for bumping the minor version is
listed in `HISTORY.md`.