Cleaned up running `mod_maker` as a script.
The `mod_maker` module had an option where if it was run directly
as a script (or main module) and --mode was specified, it did not
run through the `driver` function. I've taken this out, because
it's potentially confusing to have different behavior if `--mode`
is specified when calling this module directly versus through
`run_ginput.py`.